Add YoloV6 joined convertor for R1&R2
