+++ nginx/nginx.local.conf
@@ -6,10 +6,6 @@ upstream yolov7 {
     server yolov7:8001;
 }
 
-upstream yolov6r1 {
-    server yolov6r1:8002;
-}
-
 server {
   listen 80;
   listen [::]:80;
@@ -20,14 +16,6 @@ server {
 
   location /yolov7 {
     proxy_pass http://yolov7;
-    proxy_redirect off;
-    proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    proxy_set_header Host $http_host;
-    proxy_set_header X-Forwarded-Ssl $scheme; 
-  }
-
-  location /yolov6r1 {
-    proxy_pass http://yolov6r1;
     proxy_redirect off;
     proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
     proxy_set_header Host $http_host;
