--- yolov6r1/requirements.txt
@@ -1,25 +1,0 @@
-sanic==21.12.1
-torch==1.10.2
-
-torchvision==0.11.3
-
-onnx==1.11.0
-onnx-simplifier==0.3.6
-openvino==2021.4.2
-openvino-dev==2021.4.2
-blobconverter==1.2.9
-opencv-python>=4.1.2
-
-matplotlib>=3.2.2
-Pillow>=7.1.2
-PyYAML>=5.3.1
-requests>=2.23.0
-scipy>=1.4.1
-tqdm>=4.41.0
-
-tensorboard>=2.4.1
-
-pandas>=1.1.4
-seaborn>=0.11.0
-
-thop  # FLOPs computation
