--- yolov6r1/yolo/json/yolo.json
@@ -1,109 +1,0 @@
-{
-    "model":
-    {
-        "xml": "custom_model_name.xml",
-        "bin": "custom_model_name.bin"
-    },
-    "nn_config":
-    {
-        "output_format" : "detection",
-        "NN_family" : "YOLO",
-        "input_size": "416x416",
-        "NN_specific_metadata" :
-        {
-            "classes" : 80,
-            "coordinates" : 4,
-            "anchors" : [],
-            "anchor_masks" : {},
-            "iou_threshold" : 0.5,
-            "confidence_threshold" : 0.5
-        }
-    },
-    "mappings":
-    {
-        "labels":
-        [
-            "person",
-            "bicycle",
-            "car",
-            "motorbike",
-            "aeroplane",
-            "bus",
-            "train",
-            "truck",
-            "boat",
-            "traffic light",
-            "fire hydrant",
-            "stop sign",
-            "parking meter",
-            "bench",
-            "bird",
-            "cat",
-            "dog",
-            "horse",
-            "sheep",
-            "cow",
-            "elephant",
-            "bear",
-            "zebra",
-            "giraffe",
-            "backpack",
-            "umbrella",
-            "handbag",
-            "tie",
-            "suitcase",
-            "frisbee",
-            "skis",
-            "snowboard",
-            "sports ball",
-            "kite",
-            "baseball bat",
-            "baseball glove",
-            "skateboard",
-            "surfboard",
-            "tennis racket",
-            "bottle",
-            "wine glass",
-            "cup",
-            "fork",
-            "knife",
-            "spoon",
-            "bowl",
-            "banana",
-            "apple",
-            "sandwich",
-            "orange",
-            "broccoli",
-            "carrot",
-            "hot dog",
-            "pizza",
-            "donut",
-            "cake",
-            "chair",
-            "sofa",
-            "pottedplant",
-            "bed",
-            "diningtable",
-            "toilet",
-            "tvmonitor",
-            "laptop",
-            "mouse",
-            "remote",
-            "keyboard",
-            "cell phone",
-            "microwave",
-            "oven",
-            "toaster",
-            "sink",
-            "refrigerator",
-            "book",
-            "clock",
-            "vase",
-            "scissors",
-            "teddy bear",
-            "hair drier",
-            "toothbrush"
-        ]
-    },
-    "version": 1
-}
